moved get_git_commit_hash to logging_utils.py
